+++ handlers/user_handler.py
@@ -7,7 +7,7 @@ class UsersHandler(BaseHandler):
     @tornado.web.authenticated
     def get(self):
         username = self.get_argument('username', '')
-        users = self.session.query(User).filter(User.name.like('%{0}%'.format(username))).all()
+        users = self.session.query(User).filter(User.name.like('%{0}%'.format(username))).order_by(User.name).all()
         self.render('user/users.html', users=users)
